refactor: re-order clustering files
AI-generated. Model: Opus 5
