+++ src/Initializer/Clustering/ClusterCostModel.h
@@ -5,8 +5,8 @@
 //
 // SPDX-FileContributor: Author lists in /AUTHORS and /CITATION.cff
 
-#ifndef SEISSOL_SRC_INITIALIZER_TIMESTEPPING_CLUSTERCOSTMODEL_H_
-#define SEISSOL_SRC_INITIALIZER_TIMESTEPPING_CLUSTERCOSTMODEL_H_
+#ifndef SEISSOL_SRC_INITIALIZER_CLUSTERING_CLUSTERCOSTMODEL_H_
+#define SEISSOL_SRC_INITIALIZER_CLUSTERING_CLUSTERCOSTMODEL_H_
 
 #include <algorithm>
 #include <cstdint>
@@ -48,4 +48,4 @@ struct ClusterCostModel {
 
 } // namespace seissol::initializer
 
-#endif // SEISSOL_SRC_INITIALIZER_TIMESTEPPING_CLUSTERCOSTMODEL_H_
+#endif // SEISSOL_SRC_INITIALIZER_CLUSTERING_CLUSTERCOSTMODEL_H_
