+++ src/Initializer/Clustering/ClusterLadder.h
@@ -5,8 +5,8 @@
 //
 // SPDX-FileContributor: Author lists in /AUTHORS and /CITATION.cff
 
-#ifndef SEISSOL_SRC_INITIALIZER_TIMESTEPPING_CLUSTERLADDER_H_
-#define SEISSOL_SRC_INITIALIZER_TIMESTEPPING_CLUSTERLADDER_H_
+#ifndef SEISSOL_SRC_INITIALIZER_CLUSTERING_CLUSTERLADDER_H_
+#define SEISSOL_SRC_INITIALIZER_CLUSTERING_CLUSTERLADDER_H_
 
 #include <cstddef>
 #include <cstdint>
@@ -118,4 +118,4 @@ class ClusterLadder {
 
 } // namespace seissol::initializer
 
-#endif // SEISSOL_SRC_INITIALIZER_TIMESTEPPING_CLUSTERLADDER_H_
+#endif // SEISSOL_SRC_INITIALIZER_CLUSTERING_CLUSTERLADDER_H_
