+++ src/Initializer/Clustering/ClusterSmoother.h
@@ -7,8 +7,8 @@
 // SPDX-FileContributor: Carsten Uphoff
 // SPDX-FileContributor: Sebastian Wolf
 
-#ifndef SEISSOL_SRC_INITIALIZER_TIMESTEPPING_CLUSTERSMOOTHER_H_
-#define SEISSOL_SRC_INITIALIZER_TIMESTEPPING_CLUSTERSMOOTHER_H_
+#ifndef SEISSOL_SRC_INITIALIZER_CLUSTERING_CLUSTERSMOOTHER_H_
+#define SEISSOL_SRC_INITIALIZER_CLUSTERING_CLUSTERSMOOTHER_H_
 
 #include "Geometry/PUMLReader.h"
 #include "Initializer/BasicTypedefs.h"
@@ -89,4 +89,4 @@ class ClusterSmoother {
 
 } // namespace seissol::initializer
 
-#endif // SEISSOL_SRC_INITIALIZER_TIMESTEPPING_CLUSTERSMOOTHER_H_
+#endif // SEISSOL_SRC_INITIALIZER_CLUSTERING_CLUSTERSMOOTHER_H_
