+++ src/Initializer/Clustering/GridLadderSearch.cpp
@@ -1,0 +1,179 @@
+// SPDX-FileCopyrightText: 2026 SeisSol Group
+//
+// SPDX-License-Identifier: BSD-3-Clause
+// SPDX-LicenseComments: Full text under /LICENSE and /LICENSES/
+//
+// SPDX-FileContributor: Author lists in /AUTHORS and /CITATION.cff
+// SPDX-FileContributor: Carsten Uphoff
+// SPDX-FileContributor: Sebastian Wolf
+
+#include "Initializer/Clustering/GridLadderSearch.h"
+
+#include "Initializer/Clustering/ClusteringEvaluator.h"
+#include "Initializer/Clustering/LadderSearch.h"
+#include "Initializer/Clustering/VertexWeights/LtsWeights.h"
+#include "Initializer/Parameters/LtsParameters.h"
+
+#include <algorithm>
+#include <cassert>
+#include <cmath>
+#include <limits>
+#include <map>
+#include <optional>
+#include <utils/logger.h>
+#include <vector>
+
+namespace {
+using seissol::initializer::time_stepping::computeMaxClusterIdAfterAutoMerge;
+using seissol::initializer::time_stepping::enforceMaxClusterId;
+} // namespace
+
+namespace seissol::initializer {
+
+SearchResult GridLadderSearch::run(ClusteringEvaluator& evaluator,
+                                   const SearchConstraints& constraints) {
+  reductions_ = 0;
+
+  if (constraints.autoMergeBaseline == parameters::AutoMergeCostBaseline::BestWiggleFactor) {
+    // First compute wiggle factor without merging as baseline cost
+    logInfo() << "Using best wiggle factor as baseline cost for auto merging.";
+    logInfo() << "1. Compute best wiggle factor without merging clusters";
+    const auto baseline = sweep(evaluator, constraints, std::nullopt, false);
+    // Compute wiggle factor a second time with merging and using the previous cost as baseline
+    logInfo() << "2. Compute best wiggle factor with merging clusters, using the previous cost "
+                 "estimate as baseline";
+    return sweep(evaluator, constraints, baseline.cost, constraints.autoMerge);
+  }
+
+  assert(constraints.autoMergeBaseline == parameters::AutoMergeCostBaseline::MaxWiggleFactor);
+  return sweep(evaluator, constraints, std::nullopt, constraints.autoMerge);
+}
+
+SearchResult GridLadderSearch::sweep(ClusteringEvaluator& evaluator,
+                                     const SearchConstraints& constraints,
+                                     std::optional<double> baselineCost,
+                                     bool autoMerge) {
+  // Maps that keep track of number of clusters vs cost
+  auto mapMaxClusterIdToLowestCost = std::map<int, double>{};
+  auto maxMapClusterIdToBestWiggleFactor = std::map<int, double>{};
+
+  const double minWiggleFactor = constraints.minWiggleFactor;
+  const double maxWiggleFactor = 1.0;
+
+  const double stepSizeWiggleFactor = constraints.wiggleFactorStepsize;
+  const int numberOfStepsWiggleFactor =
+      std::ceil((maxWiggleFactor - minWiggleFactor) / stepSizeWiggleFactor) + 1;
+
+  auto computeWiggleFactor = [minWiggleFactor, stepSizeWiggleFactor, maxWiggleFactor](auto ith) {
+    return std::min(minWiggleFactor + ith * stepSizeWiggleFactor, maxWiggleFactor);
+  };
+
+  auto totalWiggleFactorReductions = 0U;
+
+  if (baselineCost) {
+    logInfo() << "Baseline cost before cluster merging is" << *baselineCost;
+  } else {
+    // Compute baselineCost cost before wiggle factor and merging of clusters
+    totalWiggleFactorReductions += evaluator.realize(maxWiggleFactor);
+    baselineCost = evaluator.globalCost(maxWiggleFactor);
+    logInfo() << "Baseline cost, without wiggle factor and cluster merging is" << *baselineCost;
+  }
+  assert(baselineCost);
+
+  const double maxAdmissibleCost = constraints.allowedPerformanceLossRatio * *baselineCost;
+
+  if (autoMerge) {
+    logInfo() << "Maximal admissible cost after cluster merging is" << maxAdmissibleCost;
+  }
+
+  for (int i = 0; i < numberOfStepsWiggleFactor; ++i) {
+    const double curWiggleFactor = computeWiggleFactor(i);
+    totalWiggleFactorReductions += evaluator.realize(curWiggleFactor);
+
+    // Note: Merging clusters does not invalidate invariance generated by enforceMaximumDifference()
+    // This can be shown by enumerating all possible cases
+    auto maxClusterIdToEnforce = constraints.maxClusterId;
+    if (autoMerge) {
+      const auto maxClusterIdAfterMerging =
+          computeMaxClusterIdAfterAutoMerge(evaluator.clusterIds(),
+                                            evaluator.cellCosts(),
+                                            evaluator.rate(),
+                                            maxAdmissibleCost,
+                                            curWiggleFactor,
+                                            evaluator.timesteps().globalMinTimeStep);
+      maxClusterIdToEnforce = std::min(maxClusterIdAfterMerging, maxClusterIdToEnforce);
+    }
+
+    evaluator.mutableClusterIds() =
+        enforceMaxClusterId(evaluator.clusterIds(), maxClusterIdToEnforce);
+    const auto maxClusterId = evaluator.globalMaxClusterId();
+
+    // Compute cost
+    const double cost = evaluator.globalCost(curWiggleFactor);
+
+    if (auto it = mapMaxClusterIdToLowestCost.find(maxClusterId);
+        it == mapMaxClusterIdToLowestCost.end() || cost <= it->second) {
+      maxMapClusterIdToBestWiggleFactor[maxClusterId] = curWiggleFactor;
+      mapMaxClusterIdToLowestCost[maxClusterId] = cost;
+    }
+  }
+
+  // Find best wiggle factor after merging of clusters
+  // We compare against cost of baselineCost.
+  int minAdmissibleMaxClusterId = std::numeric_limits<int>::max();
+  if (autoMerge) {
+    // When merging clusters, we want to find the minimum number of clusters with admissible
+    // performance.
+    bool foundAdmissibleMerge = false;
+    for (const auto& [noOfClusters, cost] : mapMaxClusterIdToLowestCost) {
+      if (cost <= maxAdmissibleCost) {
+        foundAdmissibleMerge = true;
+        minAdmissibleMaxClusterId = std::min(minAdmissibleMaxClusterId, noOfClusters);
+        logDebug() << "Admissible. cluster:" << noOfClusters << ",cost" << cost
+                   << "with wiggle factor" << maxMapClusterIdToBestWiggleFactor[noOfClusters];
+      } else {
+        logDebug() << "Not admissible. cluster:" << noOfClusters << ",cost" << cost
+                   << "with wiggle factor" << maxMapClusterIdToBestWiggleFactor[noOfClusters];
+      }
+    }
+    if (!foundAdmissibleMerge) {
+      logError() << "Found no admissible wiggle factor with cluster merge. Aborting.";
+    }
+  } else {
+    // Otherwise choose one with the smallest cost
+    minAdmissibleMaxClusterId =
+        std::min_element(mapMaxClusterIdToLowestCost.begin(),
+                         mapMaxClusterIdToLowestCost.end(),
+                         [](const auto& a, const auto& b) { return a.second < b.second; })
+            ->first;
+  }
+
+  logInfo() << "Enforcing maximum difference when finding best wiggle factor took"
+            << totalWiggleFactorReductions << "reductions.";
+  reductions_ += static_cast<int>(totalWiggleFactorReductions);
+
+  const auto bestWiggleFactor = maxMapClusterIdToBestWiggleFactor[minAdmissibleMaxClusterId];
+  const auto bestCostEstimate = mapMaxClusterIdToLowestCost[minAdmissibleMaxClusterId];
+  logInfo() << "The best wiggle factor is" << bestWiggleFactor << "with cost" << bestCostEstimate
+            << "and" << minAdmissibleMaxClusterId + 1 << "time clusters";
+
+  if (baselineCost > bestCostEstimate) {
+    logInfo() << "Cost decreased" << (*baselineCost - bestCostEstimate) / *baselineCost * 100
+              << "% with absolute cost decrease of" << *baselineCost - bestCostEstimate
+              << "compared to the baseline";
+  } else {
+    logInfo() << "Cost increased" << (bestCostEstimate - *baselineCost) / *baselineCost * 100
+              << "% with absolute cost increase of" << bestCostEstimate - *baselineCost
+              << "compared to the baseline";
+    logInfo() << "Note: Cost increased due to cluster merging!";
+  }
+
+  // The grid sweep tunes the wiggle factor only, so the ladder is the configured one -- but it
+  // is handed back expanded, because everything past the search expects a complete ladder.
+  return SearchResult{bestWiggleFactor,
+                      minAdmissibleMaxClusterId,
+                      bestCostEstimate,
+                      evaluator.configuredRatios(bestWiggleFactor)};
+}
+
+} // namespace seissol::initializer
