+++ src/Initializer/Clustering/TimestepHistogram.h
@@ -5,8 +5,8 @@
 //
 // SPDX-FileContributor: Author lists in /AUTHORS and /CITATION.cff
 
-#ifndef SEISSOL_SRC_INITIALIZER_TIMESTEPPING_TIMESTEPHISTOGRAM_H_
-#define SEISSOL_SRC_INITIALIZER_TIMESTEPPING_TIMESTEPHISTOGRAM_H_
+#ifndef SEISSOL_SRC_INITIALIZER_CLUSTERING_TIMESTEPHISTOGRAM_H_
+#define SEISSOL_SRC_INITIALIZER_CLUSTERING_TIMESTEPHISTOGRAM_H_
 
 #include <cstddef>
 #include <mpi.h>
@@ -71,4 +71,4 @@ class TimestepHistogram {
 
 } // namespace seissol::initializer
 
-#endif // SEISSOL_SRC_INITIALIZER_TIMESTEPPING_TIMESTEPHISTOGRAM_H_
+#endif // SEISSOL_SRC_INITIALIZER_CLUSTERING_TIMESTEPHISTOGRAM_H_
