+++ src/Initializer/Clustering/VertexWeights/LtsWeights.cpp
@@ -14,11 +14,12 @@
 #include "GeneratedCode/init.h"
 #include "Geometry/PUMLReader.h"
 #include "Initializer/BasicTypedefs.h"
+#include "Initializer/Clustering/ClusterLadder.h"
+#include "Initializer/Clustering/ClusterSmoother.h"
+#include "Initializer/Clustering/GridLadderSearch.h"
+#include "Initializer/Clustering/LatticeDpSearch.h"
 #include "Initializer/ParameterDB.h"
 #include "Initializer/Parameters/LtsParameters.h"
-#include "Initializer/TimeStepping/ClusterLadder.h"
-#include "Initializer/TimeStepping/ClusterSearch.h"
-#include "Initializer/TimeStepping/ClusterSmoother.h"
 #include "Initializer/TimeStepping/GlobalTimestep.h"
 #include "Parallel/MPI.h"
 #include "SeisSol.h"
