+++ src/Initializer/Clustering/VertexWeights/WeightsFactory.h
@@ -5,8 +5,8 @@
 //
 // SPDX-FileContributor: Author lists in /AUTHORS and /CITATION.cff
 
-#ifndef SEISSOL_SRC_INITIALIZER_TIMESTEPPING_LTSWEIGHTS_WEIGHTSFACTORY_H_
-#define SEISSOL_SRC_INITIALIZER_TIMESTEPPING_LTSWEIGHTS_WEIGHTSFACTORY_H_
+#ifndef SEISSOL_SRC_INITIALIZER_CLUSTERING_VERTEXWEIGHTS_WEIGHTSFACTORY_H_
+#define SEISSOL_SRC_INITIALIZER_CLUSTERING_VERTEXWEIGHTS_WEIGHTSFACTORY_H_
 
 #include "Initializer/Parameters/SeisSolParameters.h"
 #include "WeightsModels.h"
@@ -52,4 +52,4 @@ inline std::unique_ptr<LtsWeights> getLtsWeightsImplementation(parameters::LtsWe
 
 } // namespace seissol::initializer::time_stepping
 
-#endif // SEISSOL_SRC_INITIALIZER_TIMESTEPPING_LTSWEIGHTS_WEIGHTSFACTORY_H_
+#endif // SEISSOL_SRC_INITIALIZER_CLUSTERING_VERTEXWEIGHTS_WEIGHTSFACTORY_H_
