+++ src/Initializer/Clustering/VertexWeights/WeightsModels.h
@@ -5,8 +5,8 @@
 //
 // SPDX-FileContributor: Author lists in /AUTHORS and /CITATION.cff
 
-#ifndef SEISSOL_SRC_INITIALIZER_TIMESTEPPING_LTSWEIGHTS_WEIGHTSMODELS_H_
-#define SEISSOL_SRC_INITIALIZER_TIMESTEPPING_LTSWEIGHTS_WEIGHTSMODELS_H_
+#ifndef SEISSOL_SRC_INITIALIZER_CLUSTERING_VERTEXWEIGHTS_WEIGHTSMODELS_H_
+#define SEISSOL_SRC_INITIALIZER_CLUSTERING_VERTEXWEIGHTS_WEIGHTSMODELS_H_
 
 #include "LtsWeights.h"
 
@@ -54,4 +54,4 @@ class EncodedBalancedWeights : public LtsWeights {
 } // namespace initializer::time_stepping
 } // namespace seissol
 
-#endif // SEISSOL_SRC_INITIALIZER_TIMESTEPPING_LTSWEIGHTS_WEIGHTSMODELS_H_
+#endif // SEISSOL_SRC_INITIALIZER_CLUSTERING_VERTEXWEIGHTS_WEIGHTSMODELS_H_
