+++ src/Initializer/Parameters/LtsParameters.cpp
@@ -8,8 +8,8 @@
 #include "LtsParameters.h"
 
 #include "Equations/Datastructures.h"
+#include "Initializer/Clustering/ClusterCostModel.h"
 #include "Initializer/Parameters/ParameterReader.h"
-#include "Initializer/TimeStepping/ClusterCostModel.h"
 
 #include <algorithm>
 #include <cassert>
