+++ tests/Initializer/Clustering/LTSWeights.t.h
@@ -8,12 +8,12 @@
 #include "Common/Constants.h"
 #include "Geometry/PUMLReader.h"
 #include "Initializer/BasicTypedefs.h"
+#include "Initializer/Clustering/ClusterLadder.h"
+#include "Initializer/Clustering/VertexWeights/WeightsModels.h"
 #include "Initializer/FaceMap.h"
 #include "Initializer/Parameters/LtsParameters.h"
 #include "Initializer/Parameters/MeshParameters.h"
 #include "Initializer/Parameters/SeisSolParameters.h"
-#include "Initializer/TimeStepping/ClusterLadder.h"
-#include "Initializer/TimeStepping/LtsWeights/WeightsModels.h"
 #include "Initializer/Typedefs.h"
 #include "Parallel/MPI.h"
 #include "SeisSol.h"
