+++ tests/Initializer/Clustering/LadderOptimizer.t.h
@@ -5,9 +5,9 @@
 //
 // SPDX-FileContributor: Author lists in /AUTHORS and /CITATION.cff
 
-#include "Initializer/TimeStepping/ClusterCostModel.h"
-#include "Initializer/TimeStepping/LadderOptimizer.h"
-#include "Initializer/TimeStepping/TimestepHistogram.h"
+#include "Initializer/Clustering/ClusterCostModel.h"
+#include "Initializer/Clustering/LadderOptimizer.h"
+#include "Initializer/Clustering/TimestepHistogram.h"
 #include "TestHelper.h"
 
 #include <cstdint>
